refactor: configs->config; mod2->mod
