--- configs.ts
@@ -1,6 +1,0 @@
-export const DISCORD_TOKEN = Deno.env.get("DISCORD_TOKEN") ||
-  "I forgot to give a token. Blame tyler.";
-export const EVENT_HANDLER_URL = Deno.env.get("EVENT_HANDLER_URL") ||
-  "http://localhost:8000";
-export const EVENT_HANDLER_SECRET_KEY =
-  Deno.env.get("EVENT_HANDLER_SECRET_KEY") || "nick is bae";
